refactor: ヘルパーファイル名を既存パターンに合わせてリネーム
_discord-helpers.ts → discord-test-helpers.ts
