+++ spec/mcp/tools/discord-test-helpers.ts
